+++ backend/src/notion/interface/index.ts
@@ -1,0 +1,6 @@
+export * from './notion-token.interface';
+export * from './database-properties.interface';
+export * from './query-database.interface';
+export * from './get-databases.interface';
+export * from './get-database-properties.interface';
+export * from './query-database-result.interface';