+++ backend/src/snapshot/interface/index.ts
@@ -1,0 +1,4 @@
+export * from './snapshot.interface';
+export * from './query-snapshot.interface';
+export * from './query-snapshot-result.interface';
+export * from './create-query-snapshot.interface';